removed common.h

+++ ps4/include/util.h
@@ -1,10 +1,13 @@
 #ifndef Util_H
 #define Util_H
 
+#include <stdlib.h>
+#include <stdio.h>
+
 void utilStandardIORedirect(int to, int stdfd[3], fpos_t stdpos[3]);
 void utilStandardIOReset(int stdfd[3], fpos_t stdpos[3]);
 
-int utilServerCreate(uint16_t port, int backlog, int try, unsigned int sec);
+int utilServerCreate(int port, int backlog, int try, unsigned int sec);
 int utilSingleAcceptServer(int port);
 
 void *utilAllocUnsizeableFileFromDescriptor(int fd, size_t *size);
